services/email-tagger-monitor/requirements.txt
APScheduler==3.6.0
pyhumps==1.3.1
swagger-ui-py==0.2.1
quart-openapi==1.4.3
-f https://s3pypi.mettel-automation.net/igzpackages/index.html
igzpackages==2.0.1
Quart-CORS==0.1.3
